APPEALS-64051 Migrate k6 routes to API controller
